+++ src/classes/Breakpoints.ts
@@ -1,30 +1,48 @@
-import { BreakpointCollection, BreakpointToken, CreateFile } from '@/types/designTokens';
+import {
+  BreakpointCollection,
+  BreakpointToken,
+  CreateFile,
+  DesignTokensGenerator
+} from '@/types/designTokens';
 import { FigmaBreakPointComponent, FigmaFrame } from '@/types/figma';
-import { DesignTokens } from './DesignTokens.ts';
-import { getTokens, pixelate, remify, snakeCase } from '@/functions';
+import { createRootString, getTokens, pixelate, remify, snakeCase } from '@/functions';
 
-export class Breakpoints extends DesignTokens<BreakpointCollection> {
+export class Breakpoints implements DesignTokensGenerator {
+  readonly tokens: BreakpointCollection;
+
   constructor(figmaFrame: FigmaFrame) {
-    const tokens = getTokens<FigmaBreakPointComponent, BreakpointCollection, BreakpointToken>(
+    this.tokens = getTokens<FigmaBreakPointComponent, BreakpointCollection, BreakpointToken>(
       'Breakpoints',
       figmaFrame,
-      Breakpoints.getBoundingWidth
+      this.getBoundingWidth
     );
-
-    super(tokens);
   }
 
   writeTokens(createFile: CreateFile, outputDir: string, name = 'breakpoints') {
     return [createFile(name, this.tokens, outputDir, 'json')];
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   writeCssVariables(createFile: CreateFile, outputDir: string, name = 'breakpoints-vars') {
-    // TODO: Implement CSS variables for breakpoints
-    return [Promise.resolve()];
+    const { breakpointsVars } = this.generateCssBreakpointVariables(this.tokens);
+    return [createFile(name, breakpointsVars, outputDir, 'css')];
   }
 
-  static getBoundingWidth(component: FigmaBreakPointComponent): BreakpointToken | false {
+  private generateCssBreakpointVariables({ breakpoints }: BreakpointCollection) {
+    let breakpointsVars = '';
+    const breakpointsBaseName = 'breakpoint';
+
+    for (const key in breakpoints) {
+      const breakpointVarsName = `--${breakpointsBaseName}-${key}`;
+      const breakpointRemValue = `${breakpointVarsName}: ${breakpoints[key].remValue}`;
+      breakpointsVars = `${breakpointsVars}${breakpointRemValue};`;
+    }
+
+    return {
+      breakpointsVars: createRootString(breakpointsVars)
+    };
+  }
+
+  private getBoundingWidth(component: FigmaBreakPointComponent): BreakpointToken | false {
     if (!(component && component.name)) {
       return false;
     }