+++ src/classes/Spacings.ts
@@ -1,37 +1,55 @@
-import { DesignTokens } from './DesignTokens.ts';
-import { CreateFile, SpacingCollection, SpacingToken } from '@/types/designTokens';
+import {
+  CreateFile,
+  DesignTokensGenerator,
+  SpacingCollection,
+  SpacingToken
+} from '@/types/designTokens';
 import { FigmaFrame, FigmaSpacingComponent } from '@/types/figma';
-import { getTokens, pixelate, remify, snakeCase } from '@/functions';
+import { createRootString, getTokens, pixelate, remify, snakeCase } from '@/functions';
 
-export class Spacings extends DesignTokens<SpacingCollection> {
+export class Spacings implements DesignTokensGenerator {
+  readonly tokens: SpacingCollection;
+
   constructor(figmaFrame: FigmaFrame) {
-    const tokens = getTokens<FigmaSpacingComponent, SpacingCollection, SpacingToken>(
+    this.tokens = getTokens<FigmaSpacingComponent, SpacingCollection, SpacingToken>(
       'Spacings',
       figmaFrame,
-      Spacings.getBoundingWidth
+      this.getBoundingWidth
     );
-
-    super(tokens);
   }
 
   writeTokens(createFile: CreateFile, outputDir: string, name = 'spacings') {
     return [createFile(name, this.tokens, outputDir, 'json')];
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   writeCssVariables(createFile: CreateFile, outputDir: string, name = 'spacings-vars') {
-    // TODO: Implement CSS variables for spacings
-    return [Promise.resolve()];
+    const { spacingsVars } = this.generateCssSpacingVariables(this.tokens);
+    return [createFile(name, spacingsVars, outputDir, 'css')];
   }
 
-  static getBoundingWidth(component: FigmaSpacingComponent): SpacingToken | false {
+  private generateCssSpacingVariables({ spacings }: SpacingCollection) {
+    let spacingsVars = '';
+    const spacingBaseName = 'spacing';
+
+    for (const key in spacings) {
+      const spacingVarsName = `--${spacingBaseName}-${key}`;
+      const spacingRemValue = `${spacingVarsName}: ${spacings[key].remValue}`;
+      spacingsVars = `${spacingsVars}${spacingRemValue};`;
+    }
+
+    return {
+      spacingsVars: createRootString(spacingsVars)
+    };
+  }
+
+  private getBoundingWidth(component: FigmaSpacingComponent): SpacingToken | false {
     if (!(component && component.name)) {
       return false;
     }
 
     const name = snakeCase(component.name);
-    const value = pixelate(component.absoluteBoundingBox.width);
-    const remValue = remify(component.absoluteBoundingBox.width);
+    const value = pixelate(component.absoluteBoundingBox.height);
+    const remValue = remify(component.absoluteBoundingBox.height);
 
     return {
       [name]: {