+++ src/classes/index.ts
@@ -1,5 +1,6 @@
 export * from './Breakpoints.ts';
 export * from './Colors.ts';
-export * from './DesignTokens.ts';
 export * from './Spacings.ts';
 export * from './Typographies.ts';
+export * from './Borders.ts';
+export * from './Shadows.ts';