+++ src/designTokensPages.ts
@@ -1,4 +1,4 @@
-import { Breakpoints, Colors, Spacings, Typographies } from '@/classes';
+import { Breakpoints, Colors, Shadows, Spacings, Borders, Typographies } from '@/classes';
 import { DesignTokensPages } from '@/types/designTokens';
 
 const designTokensPages: DesignTokensPages[] = [
@@ -17,6 +17,14 @@ const designTokensPages: DesignTokensPages[] = [
   {
     name: 'Breakpoints',
     class: Breakpoints
+  },
+  {
+    name: 'Borders',
+    class: Borders
+  },
+  {
+    name: 'Shadows',
+    class: Shadows
   }
 ];