+++ src/functions/colorManipulation.ts
@@ -4,8 +4,18 @@ const HEX_BASE = 16;
 
 const getColor = (color: number) => Math.round(color * 255);
 
+const getAlpha = (alpha: number) => {
+  const alphaNormalize = Math.min(alpha, 1);
+  const decimals = alphaNormalize.toString().split('.')[1];
+  const alphaFixed = alphaNormalize.toFixed(2);
+  const alphaNumber = Number(alphaFixed);
+  return decimals && decimals.length > 2 ? alphaNumber : alphaNormalize;
+};
+
+const parseRgba = ({ r, g, b, a = 1 }: RgbColor) => `rgb(${r} ${g} ${b} / ${a})`;
+
 const rgbaGen = (r: number, g: number, b: number, a = 1) =>
-  `rgb(${getColor(r)} ${getColor(g)} ${getColor(b)} / ${a})`;
+  parseRgba({ r: getColor(r), g: getColor(g), b: getColor(b), a: getAlpha(a) });
 
 const rgbaGenObject = (r: number, g: number, b: number, a = 1) => ({
   r: getColor(r),
@@ -80,11 +90,6 @@ const fullColorHsl = (r: number, g: number, b: number, a = 1): HslColor => {
     l: Math.round(lightness * 100),
     a: alpha
   };
-};
-
-const parseRgba = (color: RgbColor) => {
-  const { r, g, b, a = 1 } = color;
-  return `rgb(${r} ${g} ${b} / ${a})`;
 };
 
 export { getColor, rgbaGen, rgbaGenObject, rgbToHex, fullColorHex, fullColorHsl, parseRgba };