+++ src/types/figma/index.ts
@@ -4,3 +4,5 @@ export * from './FigmaColorComponent';
 export * from './FigmaSpacingComponent';
 export * from './FigmaTypographyComponent.ts';
 export * from './FigmaBreakPointComponent.ts';
+export * from './FigmaBorderComponent.ts';
+export * from './FigmaShadowComponent.ts';